+++ Controller/AbstractDeliveryOptions.php
@@ -107,6 +107,7 @@ abstract class AbstractDeliveryOptions extends Action
         }
 
         $leadingstockmontapacking = $this->getCarrierConfig()->getLeadingStockMontapacking();
+        $disabledeliverydays = $this->getCarrierConfig()->getDisableDeliveryDays();
 
         /**
          * Retrieve Order Information
@@ -135,17 +136,19 @@ abstract class AbstractDeliveryOptions extends Action
 
         $items = $cart->getQuote()->getAllVisibleItems();
 
-
         $bAllProductsAvailable = true;
+
         foreach ($items as $item) {
 
             if ($leadingstockmontapacking) {
-                $oApi->addProduct($item->getSku(), $item->getQty(), $item->getData('length'), $item->getData('width'), $item->getData('weight')); //phpcs:ignore
+                //$oApi->addProduct($item->getSku(), $item->getQty(), $item->getData('length'), $item->getData('width'), $item->getData('weight')); //phpcs:ignore
+                $oApi->addProduct($item->getSku(), $item->getQty()); //phpcs:ignore
 
                 if (false === $oApi->checkStock($item->getSku())) {
                     $bAllProductsAvailable = false;
                     break;
                 }
+
             } else {
                 $stockItem = $item->getProduct()->getExtensionAttributes()->getStockItem();
 
@@ -155,17 +158,16 @@ abstract class AbstractDeliveryOptions extends Action
 
                 if ($stockItem->getQty() <= 0 || $stockItem->getQty() < $item->getQty()) {
 
-
                     $bAllProductsAvailable = false;
                     break;
                 }
             }
-
         }
 
-        if (false === $bAllProductsAvailable) {
+        if (false === $bAllProductsAvailable || $disabledeliverydays) {
             $oApi->setOnstock(false);
         }
+
         return $oApi;
     }
 }