+++ Model/Config/Provider/Carrier.php
@@ -11,6 +11,7 @@ class Carrier extends AbstractConfigProvider
     const XPATH_CARRIER_GOOGLEAPIKEY = 'carriers/montapacking/googleapikey';
     const XPATH_CARRIER_LOGERRORS = 'carriers/montapacking/logerrors';
     const XPATH_CARRIER_DISABLEPICKUPPOINTS = 'carriers/montapacking/disablepickuppoints';
+    const XPATH_CARRIER_DISABLEDELIVERYDAYS = 'carriers/montapacking/disabledeliverydays';
     const XPATH_CARRIER_LEADINGSTOCKMONTAPACKING = 'carriers/montapacking/leadingstockmontapacking';
     const XPATH_CARRIER_PRICE = 'carriers/montapacking/price';
 
@@ -68,6 +69,14 @@ class Carrier extends AbstractConfigProvider
     public function getDisablePickupPoints()
     {
         return $this->getConfigValue(self::XPATH_CARRIER_DISABLEPICKUPPOINTS);
+    }
+
+    /**
+     * @return string
+     */
+    public function getDisableDeliveryDays()
+    {
+        return $this->getConfigValue(self::XPATH_CARRIER_DISABLEDELIVERYDAYS);
     }
 
     /**