+++ RELEASENOTES.md
@@ -1,6 +1,6 @@
 Compatible with Open Source (CE) : 2.3, 2.4
 Stability: Stable Build
 Description:
-v1.0.9
-Add Montapacking as a shipping method to the checkout of Magento 2.
-Show available delivery Options depending on zipcode and stock stored by Montapacking
+v1.0.12
+Add Monta as a shipping method to the checkout of Magento 2.
+Show available delivery Options depending on zipcode and stock stored by Monta